added redis thing, filter for nodes

+++ alfred_rx/inc/alfredRX.h
@@ -3,14 +3,14 @@
 typedef struct bt_data_t {
     // char mac_addr[17];
     uint64_t mac_addr;
-    int32_t rssi;
+    int64_t rssi;
     time_t time;
 }bt_data_t;
 
 typedef struct bt_packed_data_t{
     uint64_t mac_addr;
     uint64_t time;
-    int16_t rssi;
+    int64_t rssi;
 
 } bt_packed_data_t __attribute__ ((packed));
